backend/requirements.txt
flask
flask-cors
supabase
google-genai
python-dotenv
pydantic
